[1.0.3] Updating release branch (#2366)

+++ featurebyte/query_graph/sql/ast/unary.py
@@ -1,9 +1,11 @@
 """
 Module for unary operations sql generation
 """
+
 from __future__ import annotations
 
-from typing import Literal, Union, cast
+from typing import Union, cast
+from typing_extensions import Literal
 
 from dataclasses import dataclass
 
@@ -111,7 +113,7 @@ class CastNode(ExpressionNode):
             expr = self.expr.sql
         type_expr = {
             "int": expressions.DataType.build("BIGINT"),
-            "float": expressions.DataType.build("FLOAT"),
+            "float": expressions.DataType.build("DOUBLE"),
             "str": expressions.DataType.build("VARCHAR"),
         }[self.new_type]
         output_expr = expressions.Cast(this=expr, to=type_expr)
